Add basic XLS exporter for projects
