+++ modules/xls_export/spec/models/xls_export/project/exporter/xls_integration_spec.rb
@@ -1,0 +1,66 @@
+require 'spec_helper'
+require 'spreadsheet'
+require 'models/projects/exporter/exportable_project_context'
+
+describe XlsExport::Project::Exporter::XLS do
+  include_context 'with a project with an arrangement of custom fields'
+  include_context 'with an instance of the described exporter'
+
+  let(:sheet) do
+    io = StringIO.new output
+    Spreadsheet.open(io).worksheets.first
+  end
+
+  let(:header) { sheet.rows.first.compact } # raw values have trailing nil
+  let(:rows) { sheet.rows.drop(1) }
+
+  describe 'empty result' do
+    before do
+      allow(instance).to receive(:records).and_return([])
+    end
+
+    it 'returns an empty XLS' do
+      expect(sheet.rows.count).to eq 1
+      expect(rows).to be_empty
+    end
+  end
+
+  it 'performs a successful export' do
+    expect(rows.count).to eq(1)
+    expect(sheet.row(1)).to eq [project.id.to_s, project.identifier, project.name, 'Off track', 'false']
+  end
+
+  describe 'custom field columns selected' do
+    before do
+      Setting.enabled_projects_columns += custom_fields.map { |cf| "cf_#{cf.id}" }
+    end
+
+    context 'when ee enabled', with_ee: %i[custom_fields_in_projects_list] do
+      it 'renders all those columns' do
+        expect(rows.count).to eq 1
+
+        cf_names = custom_fields.map(&:name)
+        expect(header).to eq ['ID', 'Identifier', 'Name', 'Status', 'Public', *cf_names]
+
+        custom_values = custom_fields.map { |cf| project.formatted_custom_value_for(cf) }
+        expect(sheet.row(1))
+          .to eq [project.id.to_s, project.identifier, project.name, 'Off track', 'false', *custom_values]
+      end
+    end
+
+    context 'when ee not enabled' do
+      it 'renders only the default columns' do
+        expect(header).to eq %w[ID Identifier Name Status Public]
+      end
+    end
+  end
+
+  context 'with no project visible' do
+    let(:current_user) { User.anonymous }
+
+    it 'does not include the project' do
+      expect(output).not_to include project.identifier
+      expect(rows).to be_empty
+    end
+  end
+end
