+++ spec/models/projects/exporter/csv_integration_spec.rb
@@ -29,69 +29,11 @@
 #++
 
 require 'spec_helper'
+require_relative 'exportable_project_context'
 
 describe Projects::Exports::CSV, 'integration', type: :model do
-  before do
-    login_as current_user
-  end
-
-  shared_let(:version_cf) { FactoryBot.create(:version_project_custom_field) }
-  shared_let(:bool_cf) { FactoryBot.create(:bool_project_custom_field) }
-  shared_let(:user_cf) { FactoryBot.create(:user_project_custom_field) }
-  shared_let(:int_cf) { FactoryBot.create(:int_project_custom_field) }
-  shared_let(:float_cf) { FactoryBot.create(:float_project_custom_field) }
-  shared_let(:text_cf) { FactoryBot.create(:text_project_custom_field) }
-  shared_let(:string_cf) { FactoryBot.create(:string_project_custom_field) }
-  shared_let(:date_cf) { FactoryBot.create(:date_project_custom_field) }
-
-  shared_let(:system_version) { FactoryBot.create(:version, sharing: 'system') }
-
-  shared_let(:role) do
-    FactoryBot.create(:role)
-  end
-
-  shared_let(:other_user) do
-    FactoryBot.create(:user,
-                      firstname: 'Other',
-                      lastname: 'User')
-  end
-
-  shared_let(:project) do
-    FactoryBot.create(:project, members: { other_user => role }).tap do |p|
-      p.send(:"custom_field_#{int_cf.id}=", 5)
-      p.send(:"custom_field_#{bool_cf.id}=", true)
-      p.send(:"custom_field_#{version_cf.id}=", system_version)
-      p.send(:"custom_field_#{float_cf.id}=", 4.5)
-      p.send(:"custom_field_#{text_cf.id}=", 'Some **long** text')
-      p.send(:"custom_field_#{string_cf.id}=", 'Some small text')
-      p.send(:"custom_field_#{date_cf.id}=", Date.today)
-      p.send(:"custom_field_#{user_cf.id}=", other_user)
-
-      p.save!(validate: false)
-    end
-  end
-
-  let(:current_user) do
-    FactoryBot.create(:user,
-                      member_in_project: project,
-                      member_with_permissions: %i(view_projects))
-  end
-  let(:query) { Queries::Projects::ProjectQuery.new }
-  let(:instance) do
-    described_class.new(query)
-  end
-
-  let(:custom_fields) { project.available_custom_fields }
-
-  let(:output) do
-    data = ''
-
-    instance.export! do |result|
-      data = result.content
-    end
-
-    data
-  end
+  include_context 'with a project with an arrangement of custom fields'
+  include_context 'with an instance of the described exporter'
 
   let(:parsed) do
     CSV.parse(output)
@@ -103,7 +45,7 @@ describe Projects::Exports::CSV, 'integration', type: :model do
 
   it 'performs a successful export' do
     expect(parsed.size).to eq(2)
-    expect(parsed.last).to eq [project.id.to_s, project.identifier, project.name, '', 'false']
+    expect(parsed.last).to eq [project.id.to_s, project.identifier, project.name, 'Off track', 'false']
   end
 
   describe 'custom field columns selected' do
@@ -120,7 +62,7 @@ describe Projects::Exports::CSV, 'integration', type: :model do
 
         custom_values = custom_fields.map { |cf| project.formatted_custom_value_for(cf) }
         expect(rows.first)
-          .to eq [project.id.to_s, project.identifier, project.name, '', 'false', *custom_values]
+          .to eq [project.id.to_s, project.identifier, project.name, 'Off track', 'false', *custom_values]
       end
     end
